SSCS-3014 Add and refactor testing classes
Because the logic to build an Event was moved into a separate service
we need to move as well the tests for that logic into a separate class.

Add a base class to reuse code for both CaseDataBuilder
and CaseDataEventBuilder classes.
